Renamed star_removal directory to StarRemoval.
To maintain consistency with directory naming convention.
