+++ GalfitModule/StarRemoval/__init__.py
