+++ GalfitModule/StarRemoval/default.conv
