+++ GalfitModule/StarRemoval/default.nnw
