+++ GalfitModule/StarRemoval/no_log_remove_stars_with_sextractor.py
