+++ GalfitModule/StarRemoval/remove_stars_with_sextractor.py
