+++ GalfitModule/StarRemoval/star_rm.param
